chore: Move to godep for until gb stabilizes
For now I'd have to use gb in combination with wgo and other
tools to make the development environment a pleasant place to be.
So until the dust settles a bit I've moved the dependency
management and vendoring back to godep.

As this also has experimental support for the vendoring solution
avilable in Go1.5 it also seems like a good long-term solution.